+++ Project/Engine/CPathManager.cpp
@@ -18,7 +18,7 @@ void CPathManager::Init()
 
 	size_t len = wcslen(m_szContentPath);
 
-	for (int i = len - 1; i > 0; i--)
+	for (size_t i = len - 1; i > 0; i--)
 	{
 		if (m_szContentPath[i] == '\\')
 		{
